requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1

# Database
sqlmodel==0.0.14
sqlite3

# GUI
tkinter
selenium==4.15.0
webdriver-manager==4.0.1

# Data Processing
pandas==2.1.3
numpy==1.24.3

# APIs
requests==2.31.0

# AI/ML
openai==0.28.1
langchain==0.0.340

# Utilities
python-dateutil==2.8.2
pydantic==2.5.0

# Graphing
yfinance==0.2.33
matplotlib==3.8.2
mplfinance==0.12.10b0